\name{EuclidijPackage-package}
\alias{EuclidijPackage-package}
\alias{EuclidijPackage}
\docType{package}
\title{
\packageTitle{EuclidijPackage}
}
\description{
\packageDescription{EuclidijPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EuclidijPackage}
\packageIndices{EuclidijPackage}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{EuclidijPackage}

Maintainer: \packageMaintainer{EuclidijPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
